Implement the changes in controls differently
The code now only changes the keyboard mapping function. It detects
if the game is in a match or in menus. When in menus, use the original
game's controls. When in a match, Z sets the shoot/charge bit every
other frame, X sets the shoot/charge bit like the old Z, C sets the
bomb bit like the old X.